requirements.txt
anyio==4.0.0
asttokens==2.4.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.2
colorama==0.4.6
executing==2.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
icecream==2.1.3
idna==3.4
mysql-connector-python==8.2.0
phpserialize==1.3
Pillow==10.1.0
protobuf==4.21.12
Pygments==2.16.1
python-slugify==8.0.1
python-telegram-bot==20.6
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
text-unidecode==1.3
urllib3==2.0.7
